style: change freemarker package.
